add M1 support
